more warning fixes

+++ engine/render/src/test/test_render.cpp
@@ -594,7 +594,7 @@ TEST_F(dmRenderTest, TextAlignment)
     int numlines;
 
     float leadings[] = { 1.0f, 2.0f, 0.5f };
-    for( int i = 0; i < sizeof(leadings)/sizeof(leadings[0]); ++i )
+    for( size_t i = 0; i < sizeof(leadings)/sizeof(leadings[0]); ++i )
     {
         float leading = leadings[i];
         tracking = 0.0f;
